requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.10.0
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-channels==0.7.0
django-cors-headers==3.8.0
django-heroku==0.3.1
django-import-export==2.5.0
django-phone-field==1.8.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
et-xmlfile==1.1.0
filelock==3.0.12
gunicorn==20.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
kiwisolver==1.3.1
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib==3.4.2
numpy==1.20.2
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
Pillow==8.3.0
platformdirs==2.0.2
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.4
service-identity==21.1.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
uritemplate==3.0.1
urllib3==1.26.5
uuid==1.30
virtualenv==20.6.0
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0